+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleServiceImp.java
@@ -16,16 +16,19 @@ public class RoleServiceImp implements RoleService{
 
     @Override
     public void add(Role role) {
+
         roleRepository.save(role);
     }
-
+    //изменён на обёртку в связи с последнем замечанием в 2.3.1
     @Override
-    public Role getById(int id) {
+    public Role getById(Long id) {
+
         return roleRepository.getById(id);
     }
 
+
     @Override
-    public Set<Role> getRoles(int [] ides) {
+    public Set<Role> getRoles(Long [] ides) {
         Set <Role> roles = new HashSet<>();
         for (int i = 0; i < ides.length; i++) {
             roles.add(roleRepository.getById(ides[i]));
